start implementing enum derive

+++ src/error.rs
@@ -20,7 +20,7 @@ pub enum Error {
     #[error("`deserialize_any` is not supported")]
     NotEnoughData,
     #[error("tag for enum is not valid")]
-    InvalidTagEncoding(usize),
+    InvalidTagEncoding(u8),
     #[error("bad response: {0}")]
     BadResponse(String),
     #[error("Unsupported column type: {0}")]
